# backend/app/api/routes/planos_trabalho.py
from fastapi import APIRouter, Depends, HTTPException, Query
from sqlalchemy.orm import Session

from backend.app.db.session import get_db
from backend.app.api.deps import check_auth
from backend.app.schemas.planos_trabalho import (
    PlanoTrabalhoCreate, PlanoTrabalhoUpdate, PlanoTrabalhoOut, PlanoTrabalhoPage
)
from backend.app.crud.planos_trabalho import crud_planos_trabalho

router = APIRouter(prefix="/planos-trabalho", tags=["Planos de Trabalho"])


@router.get("/")
def listar_planos_trabalho(
    page: int = 1,
    limit: int = 10,
    search: str | None = None,
    sort: str | None = None,
    empresa_id: str | None = Query(default=None),
    db: Session = Depends(get_db),
    current_session = Depends(check_auth),
):
    if empresa_id is None:
        empresa_id = current_session.empresas[0].empresa_id
    
    print("DEBUG empresa_id usado:", empresa_id)

    items, total = crud_planos_trabalho.listar(
        db=db,
        empresa_id=empresa_id,
        page=page,
        limit=limit,
        search=search,
        sort=sort,
    )

    return {
        "items": items,
        "total": total,
    }

@router.get("/{plano_id}", response_model=PlanoTrabalhoOut)
def getPlanoTrabalho(
    plano_id: str,
    db: Session = Depends(get_db),
    current_session = Depends(check_auth),
):
    usuario_id = current_session.usuario_id
    return crud_planos_trabalho.getPlanoTrabalho(db, usuario_id=usuario_id, plano_id=plano_id)


@router.post("/")
def criar_plano_trabalho(
    data: PlanoTrabalhoCreate,
    db: Session = Depends(get_db),
    current_session = Depends(check_auth),
):
    usuario = current_session.usuarios

    if not usuario.empresas:
        raise HTTPException(
            status_code=403,
            detail="Usuário não possui empresa vinculada"
        )

    empresa = usuario.empresas[0] 

    return crud_planos_trabalho.criar(
        db=db,
        data=data,
        empresa_id=empresa.empresa_id,
        usuario_id=usuario.usuario_id,
    )


@router.put("/{plano_id}", response_model=PlanoTrabalhoOut)
def atualizar_plano_trabalho(
    plano_id: str,
    data: PlanoTrabalhoUpdate,
    db: Session = Depends(get_db),
    current_session = Depends(check_auth),
):
    usuario_id = current_session.usuario_id
    return crud_planos_trabalho.atualizar(db, usuario_id=usuario_id, plano_id=plano_id, data=data)


@router.delete("/{plano_id}")
def deletar_plano_trabalho(
    plano_id: str,
    db: Session = Depends(get_db),
    current_session = Depends(check_auth),
):
    usuario_id = current_session.usuario_id
    crud_planos_trabalho.deletar(db, usuario_id=usuario_id, plano_id=plano_id)
    return {"ok": True}