# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XSecurity-Vault",
    "description": "Secure digital certificate management orchestrator",
    "version": "2.0.0"
  },
  "paths": {
...
    "/planos-trabalho/": {
      "get": {
        "tags": [
          "Planos de Trabalho"
        ],
        "summary": "Listar Planos Trabalho",
        "operationId": "listar_planos_trabalho_planos_trabalho__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "name": "empresa_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Empresa Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Planos de Trabalho"
        ],
        "summary": "Criar Plano Trabalho",
        "operationId": "criar_plano_trabalho_planos_trabalho__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoTrabalhoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/planos-trabalho/{plano_id}": {
      "get": {
        "tags": [
          "Planos de Trabalho"
        ],
        "summary": "Getplanotrabalho",
        "operationId": "getPlanoTrabalho_planos_trabalho__plano_id__get",
        "parameters": [
          {
            "name": "plano_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plano Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoTrabalhoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Planos de Trabalho"
        ],
        "summary": "Atualizar Plano Trabalho",
        "operationId": "atualizar_plano_trabalho_planos_trabalho__plano_id__put",
        "parameters": [
          {
            "name": "plano_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plano Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoTrabalhoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoTrabalhoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Planos de Trabalho"
        ],
        "summary": "Deletar Plano Trabalho",
        "operationId": "deletar_plano_trabalho_planos_trabalho__plano_id__delete",
        "parameters": [
          {
            "name": "plano_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plano Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ramos/": {
      "get": {
        "tags": [
          "Ramos"
        ],
        "summary": "Listar Todos Ramos",
        "description": "Lista todos os ramos de atua\u00e7\u00e3o (para selects).",
        "operationId": "listar_todos_ramos_ramos__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RamoOut"
                  },
                  "type": "array",
                  "title": "Response Listar Todos Ramos Ramos  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ramos"
        ],
        "summary": "Criar Ramo",
        "description": "Cria um novo ramo de atua\u00e7\u00e3o.",
        "operationId": "criar_ramo_ramos__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RamoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RamoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ramos/paginado": {
      "get": {
        "tags": [
          "Ramos"
        ],
        "summary": "Listar Ramos Paginado",
        "description": "Lista ramos de atua\u00e7\u00e3o com pagina\u00e7\u00e3o.",
        "operationId": "listar_ramos_paginado_ramos_paginado_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RamoOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ramos/id/{ramos_id}": {
      "get": {
        "tags": [
          "Ramos"
        ],
        "summary": "Get Ramo",
        "description": "Busca um ramo de atua\u00e7\u00e3o por ID.",
        "operationId": "get_ramo_ramos_id__ramos_id__get",
        "parameters": [
          {
            "name": "ramos_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ramos Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RamoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Ramos"
        ],
        "summary": "Atualizar Ramo",
        "description": "Atualiza um ramo de atua\u00e7\u00e3o.",
        "operationId": "atualizar_ramo_ramos_id__ramos_id__put",
        "parameters": [
          {
            "name": "ramos_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ramos Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RamoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RamoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ramos"
        ],
        "summary": "Deletar Ramo",
        "description": "Deleta um ramo de atua\u00e7\u00e3o.",
        "operationId": "deletar_ramo_ramos_id__ramos_id__delete",
        "parameters": [
          {
            "name": "ramos_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ramos Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_RamoOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RamoOut"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "title": "Page[RamoOut]"
      },
...
      "PlanoTrabalhoCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          }
        },
        "type": "object",
        "required": [
          "nome"
        ],
        "title": "PlanoTrabalhoCreate"
      },
      "PlanoTrabalhoOut": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "plano_id": {
            "type": "integer",
            "title": "Plano Id"
          },
          "empresa_id": {
            "type": "integer",
            "title": "Empresa Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "plano_id",
          "empresa_id",
          "criado_em"
        ],
        "title": "PlanoTrabalhoOut"
      },
      "PlanoTrabalhoUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          }
        },
        "type": "object",
        "title": "PlanoTrabalhoUpdate"
      },
      "RamoCreate": {
        "properties": {
          "ramo": {
            "type": "string",
            "title": "Ramo"
          }
        },
        "type": "object",
        "required": [
          "ramo"
        ],
        "title": "RamoCreate"
      },
      "RamoOut": {
        "properties": {
          "ramo": {
            "type": "string",
            "title": "Ramo"
          },
          "ramos_id": {
            "type": "integer",
            "title": "Ramos Id"
          }
        },
        "type": "object",
        "required": [
          "ramo",
          "ramos_id"
        ],
        "title": "RamoOut"
      },
      "RamoUpdate": {
        "properties": {
          "ramo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ramo"
          }
        },
        "type": "object",
        "title": "RamoUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Session token for authentication. Use the session_token from login response.",
        "scheme": "bearer"
      }
    }
  }
}